Start implementing a UI
We should be able to call cyber from the command line with the arguments it
needs to work. This is a step in that direction. Needs more work on the decode
command, to display a help message, and to support command options.

I created a package for the project that the main file farms its work out to.